Where to begin.. major upgrade.  Tilemap is not a bonefide tile map and tileset where the tileset and tilemaps are easily modified and the change propagates seamlessly into the game, no variable tweaking needed, with very flexible and robust tile system with many possibilities already integrated within..  Then I got rid of hardcoding the stop in gravity and added a collision system that works well,, but may tweak futher later, need to add the offsets to the player AnimObject qualities now to be more pixel perfect AABB collision borders instead of the default texframe size, which is wrong.  Finite State Machine work animations and works great among the physics engine setup and tilemap situation and the player all integrated into the stage system memory efficient allowing for many levels without adding memory footprint as all stages get flattened when put up on the shelf and rehydrated when needed

+++ source/misc/AnimObject.cpp
@@ -511,7 +511,7 @@ void AnimObject::render(sf::RenderWindow& tv_)
 			}
 		}
 		
-		setRect(currAnim, dir, frameIndex);
+		setRect(currAnim, dir, (int)frameIndex);
 	}
 
 	if (facingLeft && !uniDirectional && onlyRightTexture)
